+++ examples/update_file_classification.py
@@ -3,13 +3,10 @@ An example curation script to correct the session.label based on a predefined
 mapping and the classification based on the Dicom SeriesDescription data element.
 """
 
-import json
 import logging
-from pathlib import Path
 
 import flywheel
 from flywheel_gear_toolkit.utils.curator import HierarchyCurator
-from flywheel_gear_toolkit.utils.reporters import AggregatedReporter
 
 log = logging.getLogger("my_curator")
 log.setLevel("DEBUG")
@@ -18,12 +15,7 @@ log.setLevel("DEBUG")
 class Curator(HierarchyCurator):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.reporter = None
-        
-        log.info("Initiating reporter")
-        self.reporter = AggregatedReporter(
-            output_path=(Path(self.context.output_dir) / "out.csv")
-        )
+        self.config.report = True
 
     def curate_project(self, project: flywheel.Project):
         self.reporter.append_log(
@@ -31,18 +23,6 @@ class Curator(HierarchyCurator):
             label=project.label,
             msg=f"Curating files under project {project.label}",
         )
-
-    def curate_subject(self, subject: flywheel.Subject):
-        pass
-
-    def curate_session(self, session: flywheel.Session):
-        pass
-
-    def curate_acquisition(self, acquisition: flywheel.Acquisition):
-        pass
-
-    def curate_analysis(self, analysis):
-        pass
 
     def curate_file(self, file_: flywheel.FileEntry):
         try: