+++ examples/update_session_label_based_on_mapping.py
@@ -44,14 +44,9 @@ class MapLogRecord(BaseLogRecord):
 class Curator(HierarchyCurator):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
-
-        self.reporter = None
-        if self.write_report:
-            log.info("Initiating reporter")
-            self.reporter = AggregatedReporter(
-                output_path=(Path(self.context.output_dir) / "out.csv"),
-                format=MapLogRecord,  # Use custom log record
-            )
+        # Stop at session level since we don't need to curate anything under that.
+        self.config.stop_level = "session"
+        self.config.report = True
 
     def curate_session(self, session: flywheel.Session):
         log.info("Curating session %s", session.id)